requirements.txt
chromadb==0.5.0
fastapi==0.111.0
langchain==0.2.0
langchain-chroma==0.1.1
langchain-community==0.2.0
langchain-core==0.2.0
langchain-experimental==0.0.59
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
python-dotenv==1.0.1
unstructured==0.12.4